refactor: extraer lista de trails de vista
