+++ src/views/TrailsView.vue
@@ -1,13 +1,9 @@
 <script setup lang="ts">
-import TrailCard from '../components/TrailCard.vue';
 import { obtainRandomTrails } from '../services/trailsService';
 import { ref, onMounted } from 'vue';
-import { useFavoritesStore } from '../stores/favorites';
-import { RouterLink } from 'vue-router'
+import TrailList from '../components/TrailList.vue';
 
 const trails = ref();
-
-const favoritesStore = useFavoritesStore();
 
 onMounted(async () => {
   trails.value = await obtainRandomTrails(9);
@@ -20,41 +16,18 @@ onMounted(async () => {
 
 <template>
   <main>
-    <!-- Start trails container -->
     <div class="flex">
       <h1>Explore some trails</h1>
-      <div class="card__container">
-        <div v-for="(trail, i) in trails" :key="i">
-          <RouterLink :to="'/trails/' + trail.id">
-            <TrailCard 
-              :trail="trail" 
-              :isFavorite="favoritesStore.isFavorite(trail.id)"
-              @handle-favorite="favoritesStore.toggleFavorite"
-            />
-          </RouterLink>
-        </div>
+      <div class="container">
+        <TrailList :trails="trails" />
       </div>
     </div>
-    <!-- End trails container -->
   </main>
 </template>
 
 <style scoped lang="scss">
-.card__container {
-  display: grid;
-  gap: 2rem;
+
+.trail-list {
   margin: 2rem;
-
-  @media (min-width: 1200px) {
-    grid-template-columns: repeat(3, 1fr);
-  }
-
-  @media (min-width: 992px) and (max-width: 1199px) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-
-  @media (max-width: 991px) {
-    grid-template-columns: repeat(1, 1fr);
-  }
 }
 </style>
